#37 reworking on Database

+++ aaran/Common/Database/Migrations/2024_03_01_000102_create_states_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
                 $table->id();
                 $table->string('vname')->unique();
                 $table->string('state_code')->unique();
-                $table->smallInteger('active_id')->nullable();
+                $table->tinyInteger('active_id')->nullable();
             });
         }
     }
