+++ aaran/Common/Database/Migrations/2024_03_01_000114_create_despatches_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
                 $table->id();
                 $table->string('vname');
                 $table->string('vdate')->nullable();
-                $table->smallInteger('active_id')->nullable();
+                $table->tinyInteger('active_id')->nullable();
             });
         }
     }
