+++ aaran/Master/Database/Migrations/2025_01_04_000405_create_orders_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
                 $table->string('vname')->unique();
                 $table->string('order_name')->unique();
                 $table->foreignId('company_id')->references('id')->on('companies');
-                $table->smallInteger('active_id')->nullable();
+                $table->tinyInteger('active_id')->nullable();
             });
         }
     }
