+++ database/migrations/2024_01_01_000001_create_tenants_table.php
@@ -10,7 +10,7 @@ return new class extends Migration {
         Schema::create('tenants', function (Blueprint $table) {
             $table->id();
             $table->string('t_name');
-            $table->smallInteger('active_id')->nullable();
+            $table->tinyInteger('active_id')->nullable();
             $table->timestamps();
         });
     }
